backend/requirements/prod.txt
Django==4.2.10
django-cleanup==9.0.0
django-cors-headers==3.10.0
django-rest-swagger==2.2.0
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
sorl-thumbnail==12.11.0
transliterate==1.10.2
